+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added test for the `renderTo` property.
 - Updated `PHPStan` to version 2.0 and level to 10.
 - Removed separate `PHPUnit` tool.
 - Updated CI.